Closes #51, #109, #130, #129, #125, #119, #123, #112, #114, and #103

See merge request polemarch/ce!29

+++ polemarch/main/repo/tar.py
@@ -7,12 +7,18 @@ from ._base import _ArchiveRepo, shutil
 class Tar(_ArchiveRepo):
     def _extract(self, archive, path, options):
         # pylint: disable=broad-except
-        shutil.move(path, path + ".bak")
+        moved = False
+        try:
+            shutil.move(path, path + ".bak")
+            moved = True
+        except IOError:
+            pass
         try:
             with tarfile.open(archive) as arch:
                 arch.extractall(path)
         except:
             self.delete()
-            shutil.move(path + ".bak", path)
+            shutil.move(path + ".bak", path) if moved else None
+            raise
         else:
-            shutil.rmtree(path + ".bak")
+            shutil.rmtree(path + ".bak") if moved else None
